♻️ refactor: Place Test 코드 분리
