+++ server/src/test/java/com/main21/place/controller/PlaceControllerDeleteTest.java
@@ -1,0 +1,53 @@
+package com.main21.place.controller;
+
+import org.junit.jupiter.api.DisplayName;
+import org.junit.jupiter.api.Test;
+import org.mockito.Mockito;
+import org.springframework.test.web.servlet.ResultActions;
+
+import static com.main21.utils.AuthConstants.*;
+import static org.mockito.BDDMockito.given;
+import static org.mockito.Mockito.doNothing;
+import static org.springframework.restdocs.headers.HeaderDocumentation.headerWithName;
+import static org.springframework.restdocs.headers.HeaderDocumentation.requestHeaders;
+import static org.springframework.restdocs.mockmvc.MockMvcRestDocumentation.document;
+import static org.springframework.restdocs.mockmvc.RestDocumentationRequestBuilders.delete;
+import static org.springframework.restdocs.request.RequestDocumentation.parameterWithName;
+import static org.springframework.restdocs.request.RequestDocumentation.pathParameters;
+import static org.springframework.security.test.web.servlet.request.SecurityMockMvcRequestPostProcessors.csrf;
+import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.status;
+
+public class PlaceControllerDeleteTest extends PlaceControllerTest{
+
+    @Test
+    @DisplayName("DELETE 장소 삭제")
+    void deleteHosting() throws Exception{
+
+        Long placeId = 1L;
+
+        given(redisUtils.getId(Mockito.anyString())).willReturn(1L);
+        doNothing().when(placeService).deleteHosting(Mockito.anyString(), Mockito.anyLong());
+
+        ResultActions actions =
+                mockMvc.perform(
+                        delete("/place/{place-id}", placeId)
+                                .header(AUTHORIZATION, "Bearer " + accessToken)
+                                .header(REFRESH, refreshToken)
+                                .with(csrf())
+
+                );
+
+        actions
+                .andExpect(status().isOk())
+                .andDo(document("장소 삭제",
+                        requestHeaders(
+                                headerWithName(AUTHORIZATION).description("엑세스 토큰"),
+                                headerWithName(REFRESH).description("리프레시 토큰")
+                        ),
+                        pathParameters(
+                                parameterWithName("place-id").description("장소 식별자")
+                        )
+                ));
+
+    }
+}
