+++ server/src/test/java/com/main21/place/controller/PlaceControllerPatchTest.java
@@ -1,0 +1,13 @@
+package com.main21.place.controller;
+
+import org.junit.jupiter.api.DisplayName;
+import org.junit.jupiter.api.Test;
+
+public class PlaceControllerPatchTest extends PlaceControllerTest{
+
+    @Test
+    @DisplayName("PATCH 장소 수정")
+    void patchPlace() throws Exception{
+
+    }
+}
