+++ server/src/test/java/com/main21/place/controller/PlaceControllerPostTest.java
@@ -1,0 +1,14 @@
+package com.main21.place.controller;
+
+import org.junit.jupiter.api.DisplayName;
+import org.junit.jupiter.api.Test;
+
+public class PlaceControllerPostTest extends PlaceControllerTest{
+
+    @Test
+    @DisplayName("POST 장소 생성")
+    void createPlace() throws Exception{
+
+    }
+
+}
